Revert "test: fix tests"
This reverts commit 253fd5f11e2ac4b5c0d167f1ab0a443b3a1f81e1.
