+++ tests/integration/providers/s3Provider.spec.ts
@@ -6,7 +6,7 @@ import { S3Helper } from '../../helpers/s3Helper';
 import { SERVICES } from '../../../src/common/constants';
 import { getApp } from '../../../src/app';
 import { getProviderManager } from '../../../src/providers/getProvider';
-import { mockS3tS3 } from '../../helpers/mockCreator';
+import { providerConfigurationMockS32S3 } from '../../helpers/mockCreator';
 
 jest.useFakeTimers();
 
@@ -23,22 +23,23 @@ describe('S3Provider', () => {
           token: SERVICES.PROVIDER_MANAGER,
           provider: {
             useFactory: (): ProviderManager => {
-              return getProviderManager(mockS3tS3);
+              return getProviderManager(providerConfigurationMockS32S3);
             },
           },
-        },
+        }
+
       ],
     });
 
     providerManager = container.resolve(SERVICES.PROVIDER_MANAGER);
-    s3HelperSource = new S3Helper(mockS3tS3.source);
-    s3HelperDest = new S3Helper(mockS3tS3.dest);
+    s3HelperSource = new S3Helper(providerConfigurationMockS32S3.source);
+    s3HelperDest = new S3Helper(providerConfigurationMockS32S3.dest);
   });
 
   beforeEach(async () => {
     await s3HelperSource.initialize();
     await s3HelperDest.initialize();
-  });
+  })
 
   afterEach(async () => {
     await s3HelperSource.terminate();
@@ -75,7 +76,7 @@ describe('S3Provider', () => {
       const data = Buffer.from(randSentence());
 
       await providerManager.dest.postFile(`${model}/${file}`, data);
-      const result = await s3HelperDest.readFile(`${model}/${file}`);
+      const result = await s3HelperDest.readFile(providerConfigurationMockS32S3.dest.bucket, `${model}/${file}`);
 
       expect(result).toStrictEqual(data);
     });
